feat: [#154] ShareView UI 구현
